move tessla/telegraf from execute.py to separate processes
